- #151 - read data from json file - populate database from json file - get exchange rate logic from EUR

+++ CC-DTO/src/main/java/com/ubaid/ms/ccdto/ConvertedCurrency.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ConvertedCurrency {
     private Integer port;
-    private Integer convertedCurrency;
+    private Double convertedCurrency;
 }
