+++ currency-exchange-service/src/main/java/com/ubaid/ms/currencyexchangeservice/entity/ExchangeRate.java
@@ -19,7 +19,7 @@ public class ExchangeRate {
     private Date timestamp;
     private String fromCurrency;
     private String toCurrency;
-    private Integer exchangeRate;
+    private Double exchangeRate;
 
     public ExchangeValueDTO createDTO() {
         ExchangeValueDTO exchangeValueDTO = new ExchangeValueDTO();
